parquet-hadoop: bugfix dictionary page offset not set
ColumnMetaData.dictionary_page_offset is an optional field. Currently
ParquetMetadataConverter assigns a value by bare assignment.  However
optional fields needs to set a "isSet" flag as well, which is not
currently being done.  This leads to a bug that dictionary page
offsets are always read as an unset field (default to 0).

This patch changes the bare assignment of field to calling the field
setter, which internally sets the "isSet" flag as well.